Moved env and downloads to its own module
