+++ isofit/data/__init__.py
@@ -1,0 +1,3 @@
+from .ini import Ini
+
+env = Ini()
