+++ isofit/data/download.py
@@ -13,7 +13,7 @@ from zipfile import ZipFile
 
 import click
 
-from isofit.core import env
+from isofit.data import env
 
 
 def release_metadata(org, repo, tag="latest"):
@@ -178,106 +178,6 @@ tag = click.option(
 
 
 # Subcommands
-
-
-def download_data(output=None, tag="latest"):
-    """
-    Downloads the extra ISOFIT data files from the repository https://github.com/isofit/isofit-data.
-
-    Parameters
-    ----------
-    output: str | None
-        Path to output as. If None, defaults to the ini path.
-    tag: str
-        Release tag to pull from the github.
-    """
-    click.echo(f"Downloading ISOFIT data")
-
-    output = prepare_output(output, env.data)
-    if not output:
-        return
-
-    metadata = release_metadata("isofit", "isofit-data", tag)
-
-    click.echo(f"Pulling release {metadata['tag_name']}")
-    zipfile = download_file(metadata["zipball_url"], output.parent / "isofit-data.zip")
-
-    click.echo(f"Unzipping {zipfile}")
-    avail = unzip(zipfile, path=output.parent, rename=output.name)
-
-    click.echo(f"Done, now available at: {avail}")
-
-
-@cli_download.command(name="data")
-@output(help="Root directory to download data files to, ie. [path]/data")
-@tag
-def cli_data(**kwargs):
-    """\
-    Downloads the extra ISOFIT data files from the repository https://github.com/isofit/isofit-data.
-
-    Run `isofit download paths` to see default path locations.
-    There are two ways to specify output directory:
-        - `isofit --data /path/data download data`: Override the ini file. This will save the provided path for future reference.
-        - `isofit download data --output /path/data`: Temporarily set the output location. This will not be saved in the ini and may need to be manually set.
-    It is recommended to use the first style so the download path is remembered in the future.
-    """
-    download_data(**kwargs)
-
-
-def download_examples(output=None, tag="latest"):
-    """
-    Downloads the ISOFIT examples from the repository https://github.com/isofit/isofit-tutorials.
-
-    Parameters
-    ----------
-    output: str | None
-        Path to output as. If None, defaults to the ini path.
-    tag: str
-        Release tag to pull from the github.
-    """
-
-    click.echo(f"Downloading ISOFIT examples")
-
-    output = prepare_output(output, env.examples)
-    if not output:
-        return
-
-    metadata = release_metadata("isofit", "isofit-tutorials", tag)
-
-    click.echo(f"Pulling release {metadata['tag_name']}")
-    zipfile = download_file(
-        metadata["zipball_url"], output.parent / "isofit-tutorials.zip"
-    )
-
-    click.echo(f"Unzipping {zipfile}")
-    avail = unzip(zipfile, path=output.parent, rename=output.name)
-
-    click.echo(f"Done, now available at: {avail}")
-
-
-@cli_download.command(name="examples")
-@output(help="Root directory to download ISOFIT examples to, ie. [path]/examples")
-@tag
-def cli_examples(**kwargs):
-    """\
-    Downloads the ISOFIT examples from the repository https://github.com/isofit/isofit-tutorials.
-
-    Run `isofit download paths` to see default path locations.
-    There are two ways to specify output directory:
-        - `isofit --examples /path/examples download examples`: Override the ini file. This will save the provided path for future reference.
-        - `isofit download examples --output /path/examples`: Temporarily set the output location. This will not be saved in the ini and may need to be manually set.
-    It is recommended to use the first style so the download path is remembered in the future.
-    """
-    download_examples(**kwargs)
-
-
-@cli_download.command(name="all")
-def download_all():
-    """\
-    Downloads all ISOFIT extra dependencies to the locations specified in the isofit.ini file using latest tags.
-    """
-    download_data(env.data, tag="latest")
-    download_examples(env.examples, tag="latest")
 
 
 @cli_download.command(name="paths")
